[LLDB] Enable WebAssembly debugging
The idea is to use the GDB-remote protocol to connect to a Wasm engine that
implements a GDB-remote stub that offers the ability to access the engine
runtime internal state.

This patch introduce a new Process plugin wasm, with:

- class IWasmProcess that defines the interface with functions to access the
  Wasm engine state.
- class WasmProcessGDBRemote that inherits from ProcessGDBRemote and that will
  implement IWasmProcess by forwarding requests to the Wasm engine through a
  GDBRemote connection.
- class UnwindWasm that manages stack unwinding for Wasm.

Other changes:

- The GDB-remote protocol needs to be extended with a few Wasm-specific
  custom query commands, used to access Wasm-specific constructs like the Wasm
  memory, Wasm locals and globals. Changes for this are in
  GDBRemoteCommunicationClient

- DWARFExpression::Evaluate is modified to handle Wasm-specific DWARF location
  expression operators (DW_OP_WASM_location).

- Value::GetValueAsData is modified to handle read from Wasm Memory.

A new command 'wasm <port>' can be used to connect to the Wasm engine.
The latest version of V8 supports Wasm debugging when compiled with:
    v8_enable_wasm_gdb_remote_debugging = true
and run with:
    --wasm-gdb-remote